Extract CLI startup into main.ts behind a thin entry invoker
